Move lambdex to subsytem
[ci skip-rust]
[ci skip-build-wheels]
