+++ src/python/pants/backend/awslambda/python/register.py
@@ -6,10 +6,10 @@
 See https://www.pantsbuild.org/docs/awslambda-python.
 """
 
-from pants.backend.awslambda.python import lambdex
 from pants.backend.awslambda.python import rules as python_rules
 from pants.backend.awslambda.python.target_types import PythonAWSLambda
 from pants.backend.awslambda.python.target_types import rules as target_types_rules
+from pants.backend.python.subsystems import lambdex
 
 
 def rules():
