+++ src/python/pants/backend/awslambda/python/rules_test.py
@@ -9,12 +9,12 @@ from zipfile import ZipFile
 
 import pytest
 
-from pants.backend.awslambda.python.lambdex import Lambdex
-from pants.backend.awslambda.python.lambdex import rules as awslambda_python_subsystem_rules
 from pants.backend.awslambda.python.rules import PythonAwsLambdaFieldSet
 from pants.backend.awslambda.python.rules import rules as awslambda_python_rules
 from pants.backend.awslambda.python.target_types import PythonAWSLambda
 from pants.backend.awslambda.python.target_types import rules as target_rules
+from pants.backend.python.subsystems.lambdex import Lambdex
+from pants.backend.python.subsystems.lambdex import rules as awslambda_python_subsystem_rules
 from pants.backend.python.target_types import PythonLibrary
 from pants.core.goals.package import BuiltPackage
 from pants.core.target_types import Files, RelocatedFiles, Resources
